Task: Removed Calculator.py from tests and moved into Calculator directory.
