+++ Calculator/Addition.py
@@ -1,3 +1,3 @@
-def addition(a, b):
 
+def fn_addition(a, b):
     return float(a) + float(b)
