+++ Calculator/Division.py
@@ -1,4 +1,2 @@
-def divide(a, b):
-
+def fn_division(a, b):
     return float(a) / float(b)
-
