+++ Calculator/Multiplication.py
@@ -1,3 +1,2 @@
-def multiply(a, b):
-
+def fn_multiplication(a, b):
     return float(a) * float(b)
