+++ Calculator/Sqrt.py
@@ -1,7 +1,5 @@
 import math
 
 
-def sqrt(a):
-
+def fn_squareroot(a):
     return math.sqrt(float(a))
-
