+++ Calculator/Subtraction.py
@@ -1,4 +1,2 @@
-def subtraction(a, b):
-
-    return int(a) - int(b)
-
+def fn_subtraction(a, b):
+    return float(a) - float(b)
